chore: make the project easier for newcomers
* cleaned up the directory structure.
	`components` contains only visual components, which ideally
should have no side-effects (like setting the document title, or make
ajax requests) to make them easy to test and document with a story in
Storybook.
	`pages` contains "containers" components, where routing can be
configured and side-effects can be performed.
	`utils` contains non-visual components (like the route helpers)
and hooks.
* removed the routes array. Routes should be defined as React components
inside the `App`, and the main nav links should be written manually
inside the `AppLayout`.
* improved the sample pages a bit, to show some patterns that might be
useful. The Dashboard, for example, shows how to implement nested routes
with and without parameters.
* code changes to consistently use function components and named exports
(but for the page components, to allow them to be loaded lazily).
* improved the default stories to describe the "components", and not the
"pages".
* updated the App test to use react-testing-library
* favicon is now a local asset, and built and optimized automatically by
`favicons-webpack-plugin`
* added support for the `PUBLIC_URL` environment variable, to better
address #68